+++ feature/doglist/src/main/java/com/sidgowda/pawcalc/doglist/ui/DogList.kt
@@ -63,7 +63,7 @@ fun DogList(
     savedStateHandle: SavedStateHandle,
     onNavigateToOnboarding: () -> Unit,
     onNewDog: () -> Unit,
-    onEditDog: (Int) -> Unit,
+    onEditDog: (Long) -> Unit,
     onDogDetails: () -> Unit
 ) {
     val viewModel: DogListViewModel = hiltViewModel()
@@ -111,7 +111,7 @@ internal fun DogListScreen(
     modifier: Modifier = Modifier,
     viewModel: DogListViewModel,
     onNewDog: () -> Unit,
-    onEditDog: (Int) -> Unit,
+    onEditDog: (Long) -> Unit,
     onDogDetails: () -> Unit
 ) {
     LaunchedEffect(key1 = Unit) {
@@ -164,6 +164,10 @@ internal fun DogListScreen(
                             contentPadding = contentPadding,
                             state = lazyColumnState
                         ) {
+                            // Use unique id since these dogs can be added and removed
+                            // if there are 4 dogs and we remove dog of index 2 with id = 3
+                            // and add a new dog. This new dog will have id 4 and the app
+                            // will crash since there are conflicting ids.
                             items(dogListState.dogs, key = { dog -> dog.id }) { dog ->
                                 var isDogItemDismissed by remember {
                                     mutableStateOf(false)
